requirements.txt
discord.py>=2.4,<3.0
yt-dlp>=2025.1.15
python-dotenv>=1.0,<2.0
aiohttp>=3.11,<4.0
imageio-ffmpeg>=0.6,<0.7
PyNaCl>=1.5,<2.0
davey>=0.1.4,<0.2
